Tidy up CLI
Test Plan
=========

CI
