+++ Cargo.toml
@@ -79,3 +79,7 @@ opt-level = 3
 [[bench]]
 name = "benchmark"
 harness = false
+
+[[bin]]
+name = "lightswitch"
+path = "src/cli/main.rs"
